+++ x-pack/platform/test/functional/services/search_sessions.ts
